+++ client/src/main/java/org/evosuite/symbolic/vm/string/tokenizer/HasMoreTokens.java
@@ -54,7 +54,7 @@ public final class HasMoreTokens extends SymbolicFunction {
 
 		if (tokenizerExpr != null && tokenizerExpr.containsSymbolicVariable()) {
 			HasMoreTokensExpr hasMoreTokenExpr = new HasMoreTokensExpr(
-					tokenizerExpr, (long) (res ? 1L : 0L));
+					tokenizerExpr, res ? 1L : 0L);
 
 			return hasMoreTokenExpr;
